requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.3
docutils==0.22.2
ethiopian-date==1.0
filetype==1.2.0
idna==3.10
Kivy==2.3.1
Kivy-Garden==0.1.5
kivy_deps.angle==0.4.0
kivy_deps.glew==0.3.1
kivy_deps.sdl2==0.8.0
lxml==6.0.2
passlib==1.7.4
peewee==3.18.2
Pygments==2.19.2
pypiwin32==223
python-docx==1.2.0
pywin32==311
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
